Update Configuration/DataProcessing/python/Merge.py as requested by Dirk Hufnagel <Dirk.Hufnagel@cern.ch>
To support error dataset output from the
repacking in the Tier0/WMAgent I'll need
to modify the merge configuration
code in Config.DP.

New file based on CVS HEAD is here

~hufnagel/w0/releases/CMSSW_5_3_2_patch4/
  src/Configuration/DataProcessing/python/Merge.py

Change is transparent, old code interfacing
to Config.DP will not see any changes. Only
if a new parameter is passed will the
configuration we get back be different.

The only place this is used is for RAW merges
in the Tier0. But RAW merges are done with
the same release as repacking, so therefor
I need this to go into all release cycles
that are currently used or conceivable could
be used in the future for data taking.

The only somewhat urgent release I'll need is
a new 5_2_4 patch release with this change so
I can start testing against the current data.
Any other 5_2_X (or even 5_3_X) patch release
is also fine if that's easier to build.